+++ foliant/preprocessors/templateparser/engines/jinja2.py
@@ -18,6 +18,12 @@ class TemplateEngine(TemplateEngineBase):
         """Build template with jinja and return resulting string"""
         template = self._env.from_string(self._content)
         foliant_context = dict(self._context)
+        foliant_vars = dict(self._foliant_vars)
         if '_foliant_context' in self._context:
             self._context.pop('_foliant_context')
-        return template.render(_foliant_context=foliant_context, **self._context)
+        return template.render(
+            _foliant_context=foliant_context,
+            _foliant_vars=foliant_vars,
+            **self._foliant_vars,
+            **self._context
+        )