+++ foliant/preprocessors/templateparser/templateparser.py
@@ -6,17 +6,13 @@ import pkgutil
 import yaml
 import urllib
 
-import foliant.preprocessors.templateparser.engines
-
 from . import engines
 from importlib import import_module
 
 from foliant.meta.generate import load_meta, get_meta_for_chapter
 from foliant.preprocessors.utils.combined_options import (Options,
                                                           CombinedOptions,
-                                                          validate_in,
-                                                          yaml_to_dict_convertor,
-                                                          path_convertor)
+                                                          validate_in)
 from foliant.preprocessors.utils.preprocessor_ext import (BasePreprocessorExt,
                                                           allow_fail)
 
@@ -106,19 +102,32 @@ class Preprocessor(BasePreprocessorExt):
         current_pos = block.start()
         chapter = get_meta_for_chapter(self.current_filepath)
         section = chapter.get_section_by_offset(current_pos)
-        context = {'meta': section.data, 'meta_object': self.meta}
+        _foliant_vars = {
+            'meta': section.data,
+            'meta_object': self.meta,
+            'config': self.config,
+            'target': self.context['target'],
+            'backend': self.context['backend'],
+            'project_path': self.context['project_path']
+        }
+
+        context = {}
 
         # external context is loaded first, it has lowest priority
         if 'ext_context' in options:
-            context = self.load_external_context(options['ext_context'], options)
+            context.update(self.load_external_context(options['ext_context'], options))
+
         # all unrecognized params are redirected to template engine params
         context.update({p: options[p] for p in options if p not in self.tag_params})
+
         # add options from "context" param
         context.update(options.get('context', {}))
+
         template = engine(block.group('body'),
                           context,
                           options.get('engine_params', {}),
-                          self.current_filepath)
+                          self.current_filepath,
+                          _foliant_vars)
         return template.build()
 
     def apply(self):